requirements.txt
datasets == 2.6.1
elasticsearch == 7.13.0
faiss_gpu == 1.7.2
ijson == 3.1.4
ipdb == 0.13.9
jieba == 0.42.1
joblib == 1.0.1
nanopq == 0.1.10
nltk == 3.6.2
numpy == 1.19.2
pynvml == 11.4.1
requests == 2.25.1
retro_pytorch == 0.3.8
scipy == 1.7.3
spacy == 3.2.3
tqdm == 4.62.3
transformers == 4.24.0
xlrd == 2.0.1
xlwt == 1.3.0
typing
PyYAML == 6.0
